+++ src/components/homepage/stylecanvas.tsx
